Add new goal generate-no-resolve
An option to avoid resolving code at the cost of providing
the Joda-Beans version (defaulted to 2.5.0)
