+++ src/main/java/org/joda/beans/maven/JodaBeansGenerateMojo.java
@@ -15,17 +15,10 @@
  */
 package org.joda.beans.maven;
 
-import java.io.File;
-import java.util.List;
-
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Execute;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.codehaus.plexus.util.Scanner;
-import org.sonatype.plexus.build.incremental.BuildContext;
 
 /**
  * Maven plugin for generating Joda-Beans.
@@ -35,76 +28,6 @@ import org.sonatype.plexus.build.incremental.BuildContext;
         requiresDependencyResolution = ResolutionScope.COMPILE,
         threadSafe = true)
 @Execute(goal = "generate", phase = LifecyclePhase.PROCESS_SOURCES)
-public class JodaBeansGenerateMojo extends AbstractJodaBeansMojo {
-
-    private static final String[] EMPTY_STRING_ARRAY = new String[0];
-
-    //-----------------------------------------------------------------------
-    @Override
-    protected void runTool(
-            Class<?> toolClass,
-            List<String> argsList,
-            BuildContext buildContext) throws MojoExecutionException, MojoFailureException {
-
-        // only match java files that have changed according to the build context
-        // this avoids processing when there is nothing to do
-        String[] changedSourceFiles = findFiles(buildContext, getSourceDir());
-        String[] changedTestFiles = findFiles(buildContext, getTestSourceDir());
-        int sourceFilesChanged = changedSourceFiles.length;
-        int testFilesChanged = changedTestFiles.length;
-
-        // if nothing to do then exit
-        if (sourceFilesChanged == 0 && testFilesChanged == 0) {
-            logInfo("No files changed");
-            return;
-        }
-        logDebug("Files changed: main=" + sourceFilesChanged + ", test=" + testFilesChanged);
-
-        logInfo("Joda-Bean generator started, directory: " + getSourceDir() +
-                (getTestSourceDir().length() == 0 ? "" : ", test directory: " + getTestSourceDir()));
-
-        // invoke main source
-        int changedFileCount = 0;
-        if (sourceFilesChanged > 0) {
-            if (sourceFilesChanged == 1) {
-                File file = new File(getSourceDir(), changedSourceFiles[0]);
-                argsList.add(file.toString());
-                logDebug("Single file: " + argsList.get(argsList.size() - 1));
-                changedFileCount += runToolHandleChanges(toolClass, argsList, file.getParentFile(), new File(getClassesDir()));
-            } else {
-                argsList.add(getSourceDir());
-                logDebug("All files: " + argsList.get(argsList.size() - 1));
-                changedFileCount += runToolHandleChanges(toolClass, argsList, new File(getSourceDir()), new File(getClassesDir()));
-            }
-        }
-        // optionally invoke test source
-        if (testFilesChanged > 0) {
-            if (testFilesChanged == 1) {
-                File file = new File(getSourceDir(), changedTestFiles[0]);
-                argsList.set(argsList.size() - 1, file.toString());
-                logDebug("Single test file: " + argsList.get(argsList.size() - 1));
-                changedFileCount += runToolHandleChanges(toolClass, argsList, file.getParentFile(), new File(getTestClassesDir()));
-            } else {
-                argsList.set(argsList.size() - 1, getTestSourceDir());
-                logDebug("All test files: " + argsList.get(argsList.size() - 1));
-                changedFileCount += runToolHandleChanges(toolClass, argsList, new File(getTestSourceDir()), new File(getTestClassesDir()));
-            }
-        }
-
-        logInfo("Joda-Bean generator completed, " + changedFileCount + " changed files");
-    }
-
-    // find Java files
-    private String[] findFiles(BuildContext buildContext, String dirStr) {
-        File dir = new File(dirStr);
-        if (dirStr.isEmpty() || !dir.exists()) {
-            return EMPTY_STRING_ARRAY;
-        }
-        Scanner scanner = buildContext.newScanner(dir);
-        scanner.setIncludes(new String[] {"**/*.java"});
-        scanner.scan();
-        String[] changedSourceFiles = scanner.getIncludedFiles();
-        return changedSourceFiles;
-    }
+public class JodaBeansGenerateMojo extends AbstractJodaBeansGenerateMojo {
 
 }
